Rename(리뷰 내용 반영): 리뷰 내용 반영하여 함수 관리 파일 제목 및 내용 변경
